% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Address}
\alias{Address}
\title{Enterprise Apps Reseller API Objects 
Creates and manages your customers and their subscriptions.}
\usage{
Address(addressLine1 = NULL, addressLine2 = NULL, addressLine3 = NULL,
  contactName = NULL, countryCode = NULL, locality = NULL,
  organizationName = NULL, postalCode = NULL, region = NULL)
}
\arguments{
\item{addressLine1}{A customer's physical address}

\item{addressLine2}{Line 2 of the address}

\item{addressLine3}{Line 3 of the address}

\item{contactName}{The customer contact's name}

\item{countryCode}{For countryCode information, see the ISO 3166 country code elements}

\item{locality}{An example of a locality value is the city of San Francisco}

\item{organizationName}{The company or company division name}

\item{postalCode}{A postalCode example is a postal zip code such as 94043}

\item{region}{An example of a region value is CA for the state of California}
}
\value{
Address object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:09:52
filename: /Users/mark/dev/R/autoGoogleAPI/googleresellerv1.auto/R/reseller_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Address Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for address of a customer.
}
