% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{ChangePlanRequest}
\alias{ChangePlanRequest}
\title{ChangePlanRequest Object}
\usage{
ChangePlanRequest(dealCode = NULL, planName = NULL,
  purchaseOrderId = NULL, seats = NULL)
}
\arguments{
\item{dealCode}{External name of the deal code applicable for the subscription}

\item{planName}{Name of the plan to change to}

\item{purchaseOrderId}{Purchase order id for your order tracking purposes}

\item{seats}{Number/Limit of seats in the new plan}
}
\value{
ChangePlanRequest object
}
\description{
ChangePlanRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for the ChangePlan rpc request.
}
\seealso{
Other ChangePlanRequest functions: \code{\link{subscriptions.changePlan}}
}

