% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Customer}
\alias{Customer}
\title{Customer Object}
\usage{
Customer(alternateEmail = NULL, customerDomain = NULL,
  customerDomainVerified = NULL, customerId = NULL, phoneNumber = NULL,
  postalAddress = NULL, resourceUiUrl = NULL)
}
\arguments{
\item{alternateEmail}{Like the 'Customer email' in the reseller tools, this email is the secondary contact used if something happens to the customer's service such as service outage or a security issue}

\item{customerDomain}{The customer's primary domain name string}

\item{customerDomainVerified}{Whether the customer's primary domain has been verified}

\item{customerId}{This property will always be returned in a response as the unique identifier generated by Google}

\item{phoneNumber}{Customer contact phone number}

\item{postalAddress}{A customer's address information}

\item{resourceUiUrl}{URL to customer's Admin console dashboard}
}
\value{
Customer object
}
\description{
Customer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a customer.
}
\seealso{
Other Customer functions: \code{\link{customers.insert}},
  \code{\link{customers.patch}},
  \code{\link{customers.update}}
}
