% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_functions.R
\name{customers.patch}
\alias{customers.patch}
\title{Update a customer resource if one it exists and is owned by the reseller. This method supports patch semantics.}
\usage{
customers.patch(Customer, customerId)
}
\arguments{
\item{Customer}{The \link{Customer} object to pass to this method}

\item{customerId}{Id of the Customer}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/apps.order
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/apps.order)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/reseller/}{Google Documentation}

Other Customer functions: \code{\link{Customer}},
  \code{\link{customers.insert}},
  \code{\link{customers.update}}
}

