% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_functions.R
\name{otify.getwatchdetails}
\alias{otify.getwatchdetails}
\title{Returns all the details of the watch corresponding to the reseller.}
\usage{
otify.getwatchdetails()
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/apps.order
\item https://www.googleapis.com/auth/apps.order.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/apps.order, https://www.googleapis.com/auth/apps.order.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details. 

#' @importFrom googleAuthR gar_api_generator
}
\seealso{
\href{https://developers.google.com/google-apps/reseller/}{Google Documentation}
}
