% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_functions.R
\name{otify.unregister}
\alias{otify.unregister}
\title{Unregisters a Reseller for receiving notifications.}
\usage{
otify.unregister(serviceAccountEmailAddress = NULL)
}
\arguments{
\item{serviceAccountEmailAddress}{The service account which owns the Cloud-PubSub topic}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/apps.order
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/apps.order)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/reseller/}{Google Documentation}
}
