% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{RenewalSettings}
\alias{RenewalSettings}
\title{RenewalSettings Object}
\usage{
RenewalSettings(renewalType = NULL)
}
\arguments{
\item{renewalType}{Renewal settings for the annual commitment plan}
}
\value{
RenewalSettings object
}
\description{
RenewalSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a subscription renewal settings.
}
\seealso{
Other RenewalSettings functions: \code{\link{subscriptions.changeRenewalSettings}}
}
