% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_functions.R
\docType{package}
\name{reseller_googleAuthR}
\alias{reseller_googleAuthR}
\alias{reseller_googleAuthR-package}
\title{Enterprise Apps Reseller API
Creates and manages your customers and their subscriptions.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:09:52
filename: /Users/mark/dev/R/autoGoogleAPI/googleresellerv1.auto/R/reseller_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/apps.order
\item https://www.googleapis.com/auth/apps.order.readonly
}
}
