% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{ResellernotifyGetwatchdetailsResponse}
\alias{ResellernotifyGetwatchdetailsResponse}
\title{ResellernotifyGetwatchdetailsResponse Object}
\usage{
ResellernotifyGetwatchdetailsResponse(serviceAccountEmailAddresses = NULL,
  topicName = NULL)
}
\arguments{
\item{serviceAccountEmailAddresses}{List of registered service accounts}

\item{topicName}{Topic name of the PubSub}
}
\value{
ResellernotifyGetwatchdetailsResponse object
}
\description{
ResellernotifyGetwatchdetailsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for resellernotify getwatchdetails response.
}
