% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{ResellernotifyResource}
\alias{ResellernotifyResource}
\title{ResellernotifyResource Object}
\usage{
ResellernotifyResource(topicName = NULL)
}
\arguments{
\item{topicName}{Topic name of the PubSub}
}
\value{
ResellernotifyResource object
}
\description{
ResellernotifyResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for resellernotify response.
}
