% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Subscription.plan.commitmentInterval}
\alias{Subscription.plan.commitmentInterval}
\title{Subscription.plan.commitmentInterval Object}
\usage{
Subscription.plan.commitmentInterval(endTime = NULL, startTime = NULL)
}
\arguments{
\item{endTime}{An annual commitment plan's interval's endTime in milliseconds using the UNIX Epoch format}

\item{startTime}{An annual commitment plan's interval's startTime in milliseconds using UNIX Epoch format}
}
\value{
Subscription.plan.commitmentInterval object
}
\description{
Subscription.plan.commitmentInterval Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
In this version of the API, annual commitment plan's interval is one year.
}
\seealso{
Other Subscription functions: \code{\link{Subscription.plan}},
  \code{\link{Subscription.transferInfo}},
  \code{\link{Subscription.trialSettings}},
  \code{\link{Subscription}},
  \code{\link{subscriptions.insert}}
}
