% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Subscription.plan}
\alias{Subscription.plan}
\title{Subscription.plan Object}
\usage{
Subscription.plan(Subscription.plan.commitmentInterval = NULL,
  commitmentInterval = NULL, isCommitmentPlan = NULL, planName = NULL)
}
\arguments{
\item{Subscription.plan.commitmentInterval}{The \link{Subscription.plan.commitmentInterval} object or list of objects}

\item{commitmentInterval}{Interval of the commitment if it is a commitment plan}

\item{isCommitmentPlan}{Whether the plan is a commitment plan or not}

\item{planName}{The plan name of this subscription's plan}
}
\value{
Subscription.plan object
}
\description{
Subscription.plan Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Plan details of the subscription
}
\seealso{
Other Subscription functions: \code{\link{Subscription.plan.commitmentInterval}},
  \code{\link{Subscription.transferInfo}},
  \code{\link{Subscription.trialSettings}},
  \code{\link{Subscription}},
  \code{\link{subscriptions.insert}}
}

