% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Subscription}
\alias{Subscription}
\title{Subscription Object}
\usage{
Subscription(Subscription.plan = NULL,
  Subscription.plan.commitmentInterval = NULL,
  Subscription.transferInfo = NULL, Subscription.trialSettings = NULL,
  billingMethod = NULL, creationTime = NULL, customerDomain = NULL,
  customerId = NULL, dealCode = NULL, plan = NULL,
  purchaseOrderId = NULL, renewalSettings = NULL, resourceUiUrl = NULL,
  seats = NULL, skuId = NULL, status = NULL, subscriptionId = NULL,
  suspensionReasons = NULL, transferInfo = NULL, trialSettings = NULL)
}
\arguments{
\item{Subscription.plan}{The \link{Subscription.plan} object or list of objects}

\item{Subscription.plan.commitmentInterval}{The \link{Subscription.plan.commitmentInterval} object or list of objects}

\item{Subscription.transferInfo}{The \link{Subscription.transferInfo} object or list of objects}

\item{Subscription.trialSettings}{The \link{Subscription.trialSettings} object or list of objects}

\item{billingMethod}{Read-only field that returns the current billing method for a subscription}

\item{creationTime}{The creationTime property is the date when subscription was created}

\item{customerDomain}{Primary domain name of the customer}

\item{customerId}{This property will always be returned in a response as the unique identifier generated by Google}

\item{dealCode}{Google-issued code (100 char max) for discounted pricing on subscription plans}

\item{plan}{The plan property is required}

\item{purchaseOrderId}{This is an optional property}

\item{renewalSettings}{Renewal settings for the annual commitment plan}

\item{resourceUiUrl}{URL to customer's Subscriptions page in the Admin console}

\item{seats}{This is a required property}

\item{skuId}{A required property}

\item{status}{This is an optional property}

\item{subscriptionId}{The subscriptionId is the subscription identifier and is unique for each customer}

\item{suspensionReasons}{Read-only field containing an enumerable of all the current suspension reasons for a subscription}

\item{transferInfo}{Read-only transfer related information for the subscription}

\item{trialSettings}{The G Suite annual commitment and flexible payment plans can be in a 30-day free trial}
}
\value{
Subscription object
}
\description{
Subscription Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a subscription.
}
\seealso{
Other Subscription functions: \code{\link{Subscription.plan.commitmentInterval}},
  \code{\link{Subscription.plan}},
  \code{\link{Subscription.transferInfo}},
  \code{\link{Subscription.trialSettings}},
  \code{\link{subscriptions.insert}}
}
