% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Subscription.transferInfo}
\alias{Subscription.transferInfo}
\title{Subscription.transferInfo Object}
\usage{
Subscription.transferInfo(minimumTransferableSeats = NULL,
  transferabilityExpirationTime = NULL)
}
\arguments{
\item{minimumTransferableSeats}{When inserting a subscription, this is the minimum number of seats listed in the transfer order for this product}

\item{transferabilityExpirationTime}{The time when transfer token or intent to transfer will expire}
}
\value{
Subscription.transferInfo object
}
\description{
Subscription.transferInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Read-only transfer related information for the subscription. For more information, see retrieve transferable subscriptions for a customer.
}
\seealso{
Other Subscription functions: \code{\link{Subscription.plan.commitmentInterval}},
  \code{\link{Subscription.plan}},
  \code{\link{Subscription.trialSettings}},
  \code{\link{Subscription}},
  \code{\link{subscriptions.insert}}
}
