% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Subscription.trialSettings}
\alias{Subscription.trialSettings}
\title{Subscription.trialSettings Object}
\usage{
Subscription.trialSettings(isInTrial = NULL, trialEndTime = NULL)
}
\arguments{
\item{isInTrial}{Determines if a subscription's plan is in a 30-day free trial or not:}

\item{trialEndTime}{Date when the trial ends}
}
\value{
Subscription.trialSettings object
}
\description{
Subscription.trialSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The G Suite annual commitment and flexible payment plans can be in a 30-day free trial. For more information, see the API concepts.
}
\seealso{
Other Subscription functions: \code{\link{Subscription.plan.commitmentInterval}},
  \code{\link{Subscription.plan}},
  \code{\link{Subscription.transferInfo}},
  \code{\link{Subscription}},
  \code{\link{subscriptions.insert}}
}
