% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_functions.R
\name{subscriptions.insert}
\alias{subscriptions.insert}
\title{Create or transfer a subscription.}
\usage{
subscriptions.insert(Subscription, customerId, customerAuthToken = NULL)
}
\arguments{
\item{Subscription}{The \link{Subscription} object to pass to this method}

\item{customerId}{Either the customer's primary domain name or the customer's unique identifier}

\item{customerAuthToken}{The customerAuthToken query string is required when creating a resold account that transfers a direct customer's subscription or transfers another reseller customer's subscription to your reseller management}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/apps.order
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/apps.order)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/reseller/}{Google Documentation}

Other Subscription functions: \code{\link{Subscription.plan.commitmentInterval}},
  \code{\link{Subscription.plan}},
  \code{\link{Subscription.transferInfo}},
  \code{\link{Subscription.trialSettings}},
  \code{\link{Subscription}}
}
