% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_functions.R
\name{subscriptions.list}
\alias{subscriptions.list}
\title{Lists subscriptions of a reseller, optionally filtered by a customer name prefix.}
\usage{
subscriptions.list(customerAuthToken = NULL, customerId = NULL,
  customerNamePrefix = NULL, maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{customerAuthToken}{An auth token needed if the customer is not a resold customer of this reseller}

\item{customerId}{Id of the Customer}

\item{customerNamePrefix}{Prefix of the customer's domain name by which the subscriptions should be filtered}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Token to specify next page in the list}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/apps.order
\item https://www.googleapis.com/auth/apps.order.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/apps.order, https://www.googleapis.com/auth/apps.order.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/reseller/}{Google Documentation}
}

