% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Subscriptions}
\alias{Subscriptions}
\title{Subscriptions Object}
\usage{
Subscriptions(nextPageToken = NULL, subscriptions = NULL)
}
\arguments{
\item{nextPageToken}{The continuation token, used to page through large result sets}

\item{subscriptions}{The subscriptions in this page of results}
}
\value{
Subscriptions object
}
\description{
Subscriptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a subscription list.
}

