% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Address}
\alias{Address}
\title{Enterprise Apps Reseller API Objects 
Creates and manages your customers and their subscriptions.}
\usage{
Address(addressLine1 = NULL, addressLine2 = NULL, addressLine3 = NULL,
  contactName = NULL, countryCode = NULL, locality = NULL,
  organizationName = NULL, postalCode = NULL, region = NULL)
}
\arguments{
\item{addressLine1}{Address line 1 of the address}

\item{addressLine2}{Address line 2 of the address}

\item{addressLine3}{Address line 3 of the address}

\item{contactName}{Name of the contact person}

\item{countryCode}{ISO 3166 country code}

\item{locality}{Name of the locality}

\item{organizationName}{Name of the organization}

\item{postalCode}{The postal code}

\item{region}{Name of the region}
}
\value{
Address object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:27:56
filename: /Users/mark/dev/R/autoGoogleAPI/googleresellerv1.auto/R/reseller_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Address Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for address of a customer.
}

