% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{ChangePlanRequest}
\alias{ChangePlanRequest}
\title{ChangePlanRequest Object}
\usage{
ChangePlanRequest(dealCode = NULL, planName = NULL,
  purchaseOrderId = NULL, seats = NULL)
}
\arguments{
\item{dealCode}{Google-issued code (100 char max) for discounted pricing on subscription plans}

\item{planName}{The planName property is required}

\item{purchaseOrderId}{This is an optional property}

\item{seats}{This is a required property}
}
\value{
ChangePlanRequest object
}
\description{
ChangePlanRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for the ChangePlan rpc request.
}
\seealso{
Other ChangePlanRequest functions: \code{\link{subscriptions.changePlan}}
}
