% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Customer}
\alias{Customer}
\title{Customer Object}
\usage{
Customer(alternateEmail = NULL, customerDomain = NULL,
  customerDomainVerified = NULL, customerId = NULL, phoneNumber = NULL,
  postalAddress = NULL, resourceUiUrl = NULL)
}
\arguments{
\item{alternateEmail}{The alternate email of the customer}

\item{customerDomain}{The domain name of the customer}

\item{customerDomainVerified}{Whether the customer's primary domain has been verified}

\item{customerId}{The id of the customer}

\item{phoneNumber}{The phone number of the customer}

\item{postalAddress}{The postal address of the customer}

\item{resourceUiUrl}{Ui url for customer resource}
}
\value{
Customer object
}
\description{
Customer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a customer.
}
\seealso{
Other Customer functions: \code{\link{customers.insert}},
  \code{\link{customers.patch}},
  \code{\link{customers.update}}
}

