% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Subscription.plan.commitmentInterval}
\alias{Subscription.plan.commitmentInterval}
\title{Subscription.plan.commitmentInterval Object}
\usage{
Subscription.plan.commitmentInterval(endTime = NULL, startTime = NULL)
}
\arguments{
\item{endTime}{End time of the commitment interval in milliseconds since Unix epoch}

\item{startTime}{Start time of the commitment interval in milliseconds since Unix epoch}
}
\value{
Subscription.plan.commitmentInterval object
}
\description{
Subscription.plan.commitmentInterval Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Interval of the commitment if it is a commitment plan.
}
\seealso{
Other Subscription functions: \code{\link{Subscription.plan}},
  \code{\link{Subscription.transferInfo}},
  \code{\link{Subscription.trialSettings}},
  \code{\link{Subscription}},
  \code{\link{subscriptions.insert}}
}

