% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Subscription.plan}
\alias{Subscription.plan}
\title{Subscription.plan Object}
\usage{
Subscription.plan(Subscription.plan.commitmentInterval = NULL,
  commitmentInterval = NULL, isCommitmentPlan = NULL, planName = NULL)
}
\arguments{
\item{Subscription.plan.commitmentInterval}{The \link{Subscription.plan.commitmentInterval} object or list of objects}

\item{commitmentInterval}{In this version of the API, annual commitment plan's interval is one year}

\item{isCommitmentPlan}{The isCommitmentPlan property's boolean value identifies the plan as an annual commitment plan:}

\item{planName}{The planName property is required}
}
\value{
Subscription.plan object
}
\description{
Subscription.plan Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The plan property is required. In this version of the API, the G Suite plans are the flexible plan, annual commitment plan, and the 30-day free trial plan. For more information about the API's payment plans, see the API concepts.
}
\seealso{
Other Subscription functions: \code{\link{Subscription.plan.commitmentInterval}},
  \code{\link{Subscription.transferInfo}},
  \code{\link{Subscription.trialSettings}},
  \code{\link{Subscription}},
  \code{\link{subscriptions.insert}}
}
