% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Subscription}
\alias{Subscription}
\title{Subscription Object}
\usage{
Subscription(Subscription.plan = NULL,
  Subscription.plan.commitmentInterval = NULL,
  Subscription.transferInfo = NULL, Subscription.trialSettings = NULL,
  billingMethod = NULL, creationTime = NULL, customerDomain = NULL,
  customerId = NULL, dealCode = NULL, plan = NULL,
  purchaseOrderId = NULL, renewalSettings = NULL, resourceUiUrl = NULL,
  seats = NULL, skuId = NULL, status = NULL, subscriptionId = NULL,
  suspensionReasons = NULL, transferInfo = NULL, trialSettings = NULL)
}
\arguments{
\item{Subscription.plan}{The \link{Subscription.plan} object or list of objects}

\item{Subscription.plan.commitmentInterval}{The \link{Subscription.plan.commitmentInterval} object or list of objects}

\item{Subscription.transferInfo}{The \link{Subscription.transferInfo} object or list of objects}

\item{Subscription.trialSettings}{The \link{Subscription.trialSettings} object or list of objects}

\item{billingMethod}{Billing method of this subscription}

\item{creationTime}{Creation time of this subscription in milliseconds since Unix epoch}

\item{customerDomain}{Primary domain name of the customer}

\item{customerId}{The id of the customer to whom the subscription belongs}

\item{dealCode}{External name of the deal, if this subscription was provisioned under one}

\item{plan}{Plan details of the subscription}

\item{purchaseOrderId}{Purchase order id for your order tracking purposes}

\item{renewalSettings}{Renewal settings of the subscription}

\item{resourceUiUrl}{Ui url for subscription resource}

\item{seats}{Number/Limit of seats in the new plan}

\item{skuId}{Name of the sku for which this subscription is purchased}

\item{status}{Status of the subscription}

\item{subscriptionId}{The id of the subscription}

\item{suspensionReasons}{Read-only field containing an enumerable of all the current suspension reasons for a subscription}

\item{transferInfo}{Transfer related information for the subscription}

\item{trialSettings}{Trial Settings of the subscription}
}
\value{
Subscription object
}
\description{
Subscription Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a subscription.
}
\seealso{
Other Subscription functions: \code{\link{Subscription.plan.commitmentInterval}},
  \code{\link{Subscription.plan}},
  \code{\link{Subscription.transferInfo}},
  \code{\link{Subscription.trialSettings}},
  \code{\link{subscriptions.insert}}
}

