% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Subscription.transferInfo}
\alias{Subscription.transferInfo}
\title{Subscription.transferInfo Object}
\usage{
Subscription.transferInfo(minimumTransferableSeats = NULL,
  transferabilityExpirationTime = NULL)
}
\arguments{
\item{minimumTransferableSeats}{No description}

\item{transferabilityExpirationTime}{Time when transfer token or intent to transfer will expire}
}
\value{
Subscription.transferInfo object
}
\description{
Subscription.transferInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Transfer related information for the subscription.
}
\seealso{
Other Subscription functions: \code{\link{Subscription.plan.commitmentInterval}},
  \code{\link{Subscription.plan}},
  \code{\link{Subscription.trialSettings}},
  \code{\link{Subscription}},
  \code{\link{subscriptions.insert}}
}

