% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_objects.R
\name{Subscription.trialSettings}
\alias{Subscription.trialSettings}
\title{Subscription.trialSettings Object}
\usage{
Subscription.trialSettings(isInTrial = NULL, trialEndTime = NULL)
}
\arguments{
\item{isInTrial}{Whether the subscription is in trial}

\item{trialEndTime}{End time of the trial in milliseconds since Unix epoch}
}
\value{
Subscription.trialSettings object
}
\description{
Subscription.trialSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Trial Settings of the subscription.
}
\seealso{
Other Subscription functions: \code{\link{Subscription.plan.commitmentInterval}},
  \code{\link{Subscription.plan}},
  \code{\link{Subscription.transferInfo}},
  \code{\link{Subscription}},
  \code{\link{subscriptions.insert}}
}

