% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseller_functions.R
\name{subscriptions.list}
\alias{subscriptions.list}
\title{List of subscriptions managed by the reseller. The list can be all subscriptions, all of a customer's subscriptions, or all of a customer's transferable subscriptions.}
\usage{
subscriptions.list(customerAuthToken = NULL, customerId = NULL,
  customerNamePrefix = NULL, maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{customerAuthToken}{The customerAuthToken query string is required when creating a resold account that transfers a direct customer's subscription or transfers another reseller customer's subscription to your reseller management}

\item{customerId}{Either the customer's primary domain name or the customer's unique identifier}

\item{customerNamePrefix}{When retrieving all of your subscriptions and filtering for specific customers, you can enter a prefix for a customer name}

\item{maxResults}{When retrieving a large list, the maxResults is the maximum number of results per page}

\item{pageToken}{Token to specify next page in the list}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/apps.order
\item https://www.googleapis.com/auth/apps.order.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/apps.order, https://www.googleapis.com/auth/apps.order.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/reseller/}{Google Documentation}
}
